requirements.txt
pynrrd==0.4.2
